Initial Push
After around a day of getting back into coding, I have this. leaving options for encryption and decryption all eventually leading you back to the main menu and not letting you encrypt twice or decrypt a non decrypted file. a password was implemented but not tested much Im still not done